refactor(smsd): Segregation of SmsD servicer as external
