+++ lte/cloud/go/services/smsd/smsd/main.go
@@ -20,6 +20,7 @@ import (
 	"magma/lte/cloud/go/protos"
 	"magma/lte/cloud/go/services/smsd"
 	"magma/lte/cloud/go/services/smsd/servicers"
+	smsd_servicer "magma/lte/cloud/go/services/smsd/servicers/southbound"
 	storage2 "magma/lte/cloud/go/services/smsd/storage"
 	"magma/lte/cloud/go/sms_ll"
 	"magma/orc8r/cloud/go/obsidian"
@@ -49,7 +50,7 @@ func main() {
 
 	restServicer := servicers.NewRESTServicer(store)
 	obsidian.AttachHandlers(srv.EchoServer, restServicer.GetHandlers())
-	protos.RegisterSmsDServer(srv.GrpcServer, servicers.NewSMSDServicer(store, &sms_ll.DefaultSMSSerde{}))
+	protos.RegisterSmsDServer(srv.GrpcServer, smsd_servicer.NewSMSDServicer(store, &sms_ll.DefaultSMSSerde{}))
 
 	swagger_protos.RegisterSwaggerSpecServer(srv.GrpcServer, swagger.NewSpecServicerFromFile(smsd.ServiceName))
